Create archetype source branch; delete 'lite' package
This branch will be used to create the nucleus of the
pdfjt-client-archetype repo; this will consist of deleting and
relocating several sample components. It will *NOT* be merged into the
develop branch.  Once further developed, this branch will be used to
generate a Maven archetype, and the changes will be committed to the
archetype repo.

For starters, we delete the 'lite' package, which will not be used by
other client projects.